+++ src/tokens/components/stack.ts
@@ -1,0 +1,4 @@
+import stackJson from '../../../tokens/component/stack.json';
+import { flattenTokens } from '../utils';
+
+export const stackTokens = flattenTokens(stackJson as never);